parameters added

+++ routes/web.php
@@ -81,23 +81,23 @@ Route::middleware(['admin','auth'])->group(function (){
     });
     //endtestroute
     Route::get('/admin/home', 'AdminController@admin')->name('admin');
-    Route::get('/admin/samples', 'AdminController@samples')->name('samples');
-    Route::get('/admin/clients', 'AdminController@clients')->name('clients');
-    Route::get('/admin/accounts', 'AdminController@accounts')->name('accounts');
-    Route::get('/admin/parameters', 'AdminController@parameters')->name('parameters');
-    Route::post('/admin/accounts', 'AdminController@addAccount')->name('addAccount');
-    Route::delete('/admin/accounts/{accountId}', 'AdminController@destroyAccount')->name('deleteAccount');
-    Route::patch('/admin/accounts/{accountId}', 'AdminController@updateAccount')->name('updateAccount');
-    Route::post('/admin/clients', 'AdminController@addClient')->name('addClient');
-    Route::delete('/admin/clients/{clientId}', 'AdminController@destroyClient')->name('deleteClient');
-    Route::patch('/admin/clients/{clientId}', 'AdminController@updateClient')->name('updateClient');
-    Route::post('/admin/samples', 'AdminController@addSample')->name('addSample');
-    Route::post('/admin/parameters', 'AdminController@addParameter')->name('addParameter');
-    Route::delete('/admin/clients/{parameterId}', 'AdminController@destroyParameter')->name('deleteParameter');
-    Route::patch('/admin/clients/{parameterId}', 'AdminController@updateParameter')->name('updateParameter');
+    Route::get('/admin/samples', 'AdminController@samples')->name('samples-admin');
+    Route::get('/admin/clients', 'AdminController@clients')->name('clients-admin');
+    Route::get('/admin/accounts', 'AdminController@accounts')->name('accounts-admin');
+    Route::get('/admin/parameters', 'AdminController@parameters')->name('parameters-admin');
+    Route::post('/admin/accounts', 'AdminController@addAccount')->name('addAccount-admin');
+    Route::delete('/admin/accounts/{accountId}', 'AdminController@destroyAccount')->name('deleteAccount-admin');
+    Route::patch('/admin/accounts/{accountId}', 'AdminController@updateAccount')->name('updateAccount-admin');
+    Route::post('/admin/clients', 'AdminController@addClient')->name('addClient-admin');
+    Route::delete('/admin/clients/{clientId}', 'AdminController@destroyClient')->name('deleteClient-admin');
+    Route::patch('/admin/clients/{clientId}', 'AdminController@updateClient')->name('updateClient-admin');
+    Route::post('/admin/samples', 'AdminController@addSample')->name('addSample-admin');
+    Route::post('/admin/parameters', 'AdminController@addParameter')->name('addParameter-admin');
+    Route::delete('/admin/parameters/{parameterId}', 'AdminController@destroyParameter')->name('deleteParameter-admin');
+    Route::patch('/admin/parameters/{parameterId}', 'AdminController@updateParameter')->name('updateParameter-admin');
     //Inventory routes
-    Route::get('/admin/inventory/chemicals', 'AdminController@chemicals')->name('inventory-chemicals');
-    Route::get('/admin/inventory/glassware', 'AdminController@glassware')->name('inventory-glassware');
+    Route::get('/admin/inventory/chemicals', 'AdminController@chemicals')->name('inventory-chemicals-admin');
+    Route::get('/admin/inventory/glassware', 'AdminController@glassware')->name('inventory-glassware-admin');
 });
 // END ADMIN ROUTES
